refactor paramsets/paramview into parameters
