+++ src/pyhf/modifiers/normsys.py
@@ -1,10 +1,9 @@
 import logging
 
 from . import modifier
-from ..paramsets import constrained_by_normal
 from .. import get_backend, events
 from .. import interpolators
-from ..paramview import ParamViewer
+from ..parameters import constrained_by_normal, ParamViewer
 
 log = logging.getLogger(__name__)
