+++ src/pyhf/modifiers/shapesys.py
@@ -1,9 +1,8 @@
 import logging
 
 from . import modifier
-from ..paramsets import constrained_by_poisson
 from .. import get_backend, default_backend, events
-from ..paramview import ParamViewer
+from ..parameters import constrained_by_poisson, ParamViewer
 
 log = logging.getLogger(__name__)
