+++ src/pyhf/parameters/__init__.py
@@ -1,0 +1,21 @@
+import logging
+
+log = logging.getLogger(__name__)
+
+from .paramsets import (
+    paramset,
+    unconstrained,
+    constrained_by_normal,
+    constrained_by_poisson,
+)
+from . import utils
+from .paramview import ParamViewer
+
+__all__ = [
+    'paramset',
+    'unconstrained',
+    'constrained_by_normal',
+    'constrained_by_poisson',
+    'utils',
+    'ParamViewer',
+]
